+++ examples/src/tower/client.rs
@@ -15,7 +15,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let channel = ServiceBuilder::new()
         // Interceptors can be also be applied as middleware
-        .layer(tonic::service::interceptor_fn(intercept))
+        .layer(tonic::service::interceptor(intercept))
         .layer_fn(AuthSvc::new)
         .service(channel);
 
@@ -61,6 +61,7 @@ mod service {
     impl Service<Request<BoxBody>> for AuthSvc {
         type Response = Response<Body>;
         type Error = Box<dyn std::error::Error + Send + Sync>;
+        #[allow(clippy::type_complexity)]
         type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
 
         fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {