+++ tonic/src/request.rs
@@ -195,7 +195,7 @@ impl<T> Request<T> {
         Request {
             metadata: self.metadata,
             message,
-            extensions: Extensions::new(),
+            extensions: self.extensions,
         }
     }
 
@@ -300,13 +300,13 @@ impl<T> Request<T> {
     /// Extensions can be set in interceptors:
     ///
     /// ```no_run
-    /// use tonic::{Request, service::interceptor_fn};
+    /// use tonic::{Request, service::interceptor};
     ///
     /// struct MyExtension {
     ///     some_piece_of_data: String,
     /// }
     ///
-    /// interceptor_fn(|mut request: Request<()>| {
+    /// interceptor(|mut request: Request<()>| {
     ///     request.extensions_mut().insert(MyExtension {
     ///         some_piece_of_data: "foo".to_string(),
     ///     });